+++ src/mixer/track_id.rs
@@ -1,2 +1,2 @@
 #[derive(Clone, Copy, Default, Eq, Hash, PartialEq, Debug)]
-pub struct TrackID(pub usize);
+pub struct TrackID(pub u64);
