+++ src/track/region_id.rs
@@ -1,4 +1,4 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Default, Eq, Hash, PartialEq, Debug, Serialize, Deserialize)]
-pub struct RegionID(pub usize);
+pub struct RegionID(pub u64);
